Refactor site details page, builds component for site details
