+++ src/client/components/site-details/siteDetails.module.less
@@ -1,3 +1,7 @@
 .centerText {
     text-align: center;
 }
+
+.heading {
+    margin-top: 1rem;
+}
